refactore: replace stations from admin page to shared components
