+++ src/app/app.module.ts
@@ -8,14 +8,11 @@ import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCardModule } from '@angular/material/card';
-import { LeafletModule } from '@asymmetrik/ngx-leaflet';
-import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatTableModule } from '@angular/material/table';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatDividerModule } from '@angular/material/divider';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -68,17 +65,13 @@ import { UserEffects } from './core/store/user/user.effects';
     StoreDevtoolsModule.instrument(),
     UserModule,
     MatButtonModule,
-    MatIconModule,
-    LeafletModule,
     MatCardModule,
-    MatPaginatorModule,
     MatTableModule,
     MatFormFieldModule,
     MatInputModule,
     MatIconModule,
     MatSelectModule,
     ReactiveFormsModule,
-    MatProgressSpinnerModule,
     MatDividerModule,
     SharedModule,
   ],
